include jp_free/parse and error strings in library
