+++ matcher.h
@@ -26,12 +26,5 @@
 	#include <json-c/json.h>
 #endif
 
-#include "ast.h"
-
-typedef void (*jp_match_cb_t)(struct json_object *res, void *priv);
-
-struct json_object *
-jp_match(struct jp_opcode *path, struct json_object *jsobj,
-         jp_match_cb_t cb, void *priv);
 
 #endif
